+++ src/models/BikeSystemModel/Bikes.ts
@@ -2,9 +2,9 @@ import mongoose, { Document, Schema } from "mongoose";
 
 // Variant interface for multiple options
 export interface IBikeVariant {
-  name: string; // "Standard", "Multi-optional"
+  name: string;
   features: string[];
-  priceAdjustment: number; // Additional price from base
+  priceAdjustment: number;
   isAvailable: boolean;
 }
 
@@ -13,10 +13,10 @@ export interface IPriceBreakdown {
   exShowroomPrice: number;
   rtoCharges: number;
   insuranceComprehensive: number;
-  onRoadPrice?: number; // Auto-calculated field
+  onRoadPrice?: number;
 }
 
-// Enhanced Bikes Document interface
+// Enhanced Bikes Document interface (without images array)
 export interface IBikesDocument extends Document {
   modelName: string;
   mainCategory: "bike" | "scooter";
@@ -31,38 +31,18 @@ export interface IBikesDocument extends Document {
     | "automatic"
     | "gearless";
   year: number;
-
-  // New variant system
   variants: IBikeVariant[];
-
-  // Enhanced pricing structure
   priceBreakdown: IPriceBreakdown;
-
-  // Technical specs
   engineSize: string;
   power: number;
   transmission: string;
   fuelNorms: "BS4" | "BS6" | "BS6 Phase 2" | "Electric";
   isE20Efficiency: boolean;
-
-  // Features and appearance
   features: string[];
   colors: string[];
-
-  // Enhanced image system for multiple photos
-  images: {
-    src: string;
-    alt: string;
-    cloudinaryPublicId: string;
-    isPrimary: boolean;
-  }[];
-
-  // Stock and availability
   stockAvailable: number;
   isNewModel?: boolean;
   isActive: boolean;
-
-  // Key specifications for display
   keySpecifications: {
     engine?: string;
     power?: string;
@@ -71,7 +51,6 @@ export interface IBikesDocument extends Document {
     fuelNorms?: string;
     isE20Efficiency?: boolean;
   };
-
   createdAt: Date;
   updatedAt: Date;
 }
@@ -111,27 +90,6 @@ const PriceBreakdownSchema = new Schema<IPriceBreakdown>({
   },
   onRoadPrice: {
     type: Number,
-    // This will be auto-calculated
-  },
-});
-
-// Enhanced image schema for multiple photos
-const BikeImageSchema = new Schema({
-  src: {
-    type: String,
-    required: [true, "Image URL is required"],
-  },
-  alt: {
-    type: String,
-    required: [true, "Image alt text is required"],
-  },
-  cloudinaryPublicId: {
-    type: String,
-    required: [true, "Cloudinary public ID is required"],
-  },
-  isPrimary: {
-    type: Boolean,
-    default: false,
   },
 });
 
@@ -143,7 +101,6 @@ const BikesSchema = new Schema<IBikesDocument>(
       trim: true,
       maxlength: [100, "Model name cannot exceed 100 characters"],
     },
-
     mainCategory: {
       type: String,
       required: [true, "Please specify main category"],
@@ -152,7 +109,6 @@ const BikesSchema = new Schema<IBikesDocument>(
         message: "Main category must be either bike or scooter",
       },
     },
-
     category: {
       type: String,
       required: [true, "Please add vehicle category"],
@@ -171,7 +127,6 @@ const BikesSchema = new Schema<IBikesDocument>(
         message: "Invalid category selected",
       },
     },
-
     year: {
       type: Number,
       required: [true, "Please add manufacturing year"],
@@ -181,8 +136,6 @@ const BikesSchema = new Schema<IBikesDocument>(
         "Year cannot be more than 2 years in future",
       ],
     },
-
-    // New variant system
     variants: {
       type: [BikeVariantSchema],
       validate: {
@@ -192,31 +145,25 @@ const BikesSchema = new Schema<IBikesDocument>(
         message: "At least one variant is required",
       },
     },
-
-    // Enhanced pricing structure
     priceBreakdown: {
       type: PriceBreakdownSchema,
       required: [true, "Price breakdown is required"],
     },
-
     engineSize: {
       type: String,
       required: [true, "Please add engine details"],
       trim: true,
     },
-
     power: {
       type: Number,
       required: [true, "Please add power specifications"],
       min: [1, "Power must be positive"],
     },
-
     transmission: {
       type: String,
       required: [true, "Please add transmission details"],
       trim: true,
     },
-
     fuelNorms: {
       type: String,
       required: [true, "Please specify fuel norms"],
@@ -226,18 +173,15 @@ const BikesSchema = new Schema<IBikesDocument>(
           "Invalid fuel norm. Must be BS4, BS6, BS6 Phase 2, or Electric",
       },
     },
-
     isE20Efficiency: {
       type: Boolean,
       default: false,
       required: [true, "Please specify E20 efficiency compatibility"],
     },
-
     features: {
       type: [String],
       default: [],
     },
-
     colors: {
       type: [String],
       validate: {
@@ -247,35 +191,19 @@ const BikesSchema = new Schema<IBikesDocument>(
         message: "At least one color option is required",
       },
     },
-
-    // Enhanced image system
-    images: {
-      type: [BikeImageSchema],
-      validate: {
-        validator: function (images: any[]) {
-          return images && images.length > 0;
-        },
-        message: "At least one image is required",
-      },
-    },
-
     stockAvailable: {
       type: Number,
       default: 0,
       min: [0, "Stock cannot be negative"],
     },
-
     isNewModel: {
       type: Boolean,
       default: false,
     },
-
     isActive: {
       type: Boolean,
       default: true,
     },
-
-    // Key specifications for quick display
     keySpecifications: {
       engine: String,
       power: String,
@@ -312,24 +240,7 @@ BikesSchema.pre("save", function (next) {
   next();
 });
 
-// Ensure only one primary image
-BikesSchema.pre("save", function (next) {
-  if (this.images && this.images.length > 0) {
-    const primaryImages = this.images.filter((img) => img.isPrimary);
-    if (primaryImages.length === 0) {
-      // Set first image as primary if none is set
-      this.images[0].isPrimary = true;
-    } else if (primaryImages.length > 1) {
-      // Keep only the first primary image
-      this.images.forEach((img, index) => {
-        img.isPrimary = index === 0;
-      });
-    }
-  }
-  next();
-});
-
-// Create compound unique index for modelName and year to prevent duplicates
+// Create compound unique index
 BikesSchema.index({ modelName: 1, year: 1 }, { unique: true });
 
 // Additional performance indexes
@@ -342,5 +253,4 @@ BikesSchema.index({ isE20Efficiency: 1 });
 BikesSchema.index({ "priceBreakdown.onRoadPrice": 1 });
 
 const BikeModel = mongoose.model<IBikesDocument>("Bikes", BikesSchema);
-
 export default BikeModel;
